fix: ajustando os diretórios dos hooks
