+++ src/pages/ImportacaoDados/Habilitados/hooks/usePostImportacaoArquivosHabilitados.tsx
@@ -1,30 +1,23 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { App } from "antd";
-import { API } from "../services";
-import type { IImportacaoVagasPayload } from "../pages/ImportacaoDados/Vagas/hooks/types";
+import { API } from "../../../../services";
+import type { IImportacaoHabilitadosPayload } from "./types";
 
-type UsePostImportacaoArquivosVagasOptions = {
-  onSuccess?: () => void;
-};
-
-export const usePostImportacaoArquivosVagas = (
-  options?: UsePostImportacaoArquivosVagasOptions
-) => {
+export const usePostImportacaoArquivosHabilitados = () => {
   const queryClient = useQueryClient();
   const { notification } = App.useApp();
 
   return useMutation({
-    mutationFn: (payload: IImportacaoVagasPayload) =>
-      API.ImportacaoDados.postImportacaoArquivosVagas(payload).response,
+    mutationFn: (payload: IImportacaoHabilitadosPayload) =>
+      API.ImportacaoDados.postImportacaoArquivosHabilitados(payload).response,
     onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ["getImportacaoArquivosVagas"] });
+      queryClient.invalidateQueries({ queryKey: ["getImportacaoArquivosHabilitados"] });
       notification.success({
         message: "Importação Realizada",
         description: "A importação dos dados foi processada com sucesso!",
         placement: "top",
         duration: 3.5,
       });
-      options?.onSuccess?.();
     },
     onError: () => {
       notification.error({
